refactor(usePosterDownloader): dom을 이미지로 만드는 로직 hook으로 분리
에디터뿐만 아니라 결과 페이지에서도 다운로드 진행해야 하기 때문에 분리해서 재사용성 높임
