+++ card-capture/src/components/editor/EditingArea/views/CardArea.tsx
@@ -8,15 +8,14 @@ import LayerAddBox from '@/components/editor/EditingArea/views/LayerAddBox';
 import Button from '@/components/common/Button/Button';
 import { useFocusStore } from '@/store/useFocusStore';
 import ImageBox from '@/components/editor/EditingArea/components/ImageBox/ImageBox';
-import { Dialog, DialogContent, DialogDescription, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
-import React, { useCallback, useEffect, useRef, useState } from 'react';
+import { Dialog, DialogContent, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
+import React, { useEffect, useState } from 'react';
 import { useCardsStore } from '@/store/useCardsStore';
-import { saveAs } from 'file-saver';
-import { toPng } from 'html-to-image';
-import { Progress } from '@/components/ui/progress';
 import { jsonUtils } from '@/utils';
 import IllustBox from '@/components/editor/EditingArea/components/IllustBox/IllustBox';
 import useAmplitudeContext from '@/hooks/useAmplitudeContext';
+import usePosterDownloader from '@/hooks/usePosterDownloader';
+import DownloadProgressModal from '@/components/common/Progress/DownloadProgressModal';
 
 const CardArea = ({ card }: { card: Card }) => {
   const cardId = card.id;
@@ -64,65 +63,10 @@ const CardArea = ({ card }: { card: Card }) => {
   };
 
   /**
-   * card에 그려진 dom을 image export 하는 handler
-   * html-to-image와 file-saver 사용
+   * card에 그려진 dom을 image export 하는 hook
+   * Ref에 그려진 요소들을 이미지로 변환, 다운로드 함
    */
-  const [isDownloading, setIsDownloading] = useState(false);
-  const [progress, setProgress] = useState(10);
-  const cardRef = useRef<HTMLDivElement>(null);
-
-  const downloadCardHandler = async () => {
-    if (!cardRef || !cardRef.current) return;
-
-    // 포커스 된 항목이 있는 경우 FocusBox의 테두리도 함께 출력될 수 있으므로 제거
-    setFocusedLayerId(-1);
-    await new Promise(resolve => setTimeout(resolve, 100));
-
-    // 다운로드가 시작되면서 대기 모달이 떠야 함으로 상태 변경
-    setIsDownloading(true);
-
-    try {
-      const cardArea = cardRef.current;
-
-      // html-to-image 사용하여 이미지로 변경
-      const dataUrl = await toPng(cardArea, {
-        includeQueryParams: true, // 이미지 src 하나로만 나오는 오류 해결
-        quality: 0.8, // 품질 설정
-        width: cardArea.offsetWidth,
-        height: cardArea.offsetHeight,
-      });
-
-      // file-saver 라이브러리 사용하여 자동 다운로드
-      saveAs(dataUrl, 'card-capture-image.png');
-    } catch (e) {
-      console.error('Error converting card to image', e);
-    } finally {
-      setIsDownloading(false);
-      setProgress(10);
-    }
-  };
-
-  /**
-   * progress 바를 10씩 증가시키는 함수
-   */
-  const increaseProgress = useCallback(() => {
-    if (progress < 90) {
-      setProgress(prevProgress => prevProgress + 10);
-    }
-  }, [progress]);
-
-  /**
-   * 3초에 한번씩 progress를 증가시키는 로직
-   */
-  useEffect(() => {
-    let timer: NodeJS.Timeout | undefined;
-    if (isDownloading) {
-      timer = setInterval(increaseProgress, 300);
-    }
-    return () => {
-      if (timer) clearInterval(timer);
-    };
-  }, [isDownloading, increaseProgress]);
+  const { cardRef, isDownloading, setIsDownloading, downloadCardHandler } = usePosterDownloader();
 
   /**
    * 에디터 페이지에서 버튼 클릭에 대한 tracking
@@ -182,18 +126,7 @@ const CardArea = ({ card }: { card: Card }) => {
       </div>
 
       {/* export 중에 뜨는 모달 */}
-      <Dialog open={isDownloading} onOpenChange={setIsDownloading}>
-        <DialogContent className="flex w-[400px] flex-col items-center justify-center px-[20px] py-[40px]">
-          <div className="flex flex-col items-center justify-center gap-1">
-            <DialogTitle>카드뉴스를 이미지로 변환하고 있어요!</DialogTitle>
-            <DialogDescription>대기 창을 닫아도 다운로드는 문제없이 진행됩니다</DialogDescription>
-          </div>
-          <div className="flex flex-col">
-            <Progress value={progress} className="w-[250px]" />
-            <div className="h-3 w-3 animate-spin" />
-          </div>
-        </DialogContent>
-      </Dialog>
+      <DownloadProgressModal isOpen={isDownloading} onOpenChange={setIsDownloading} />
 
       {/* card에 요소 출력 */}
       <div
